updating progress
